todos: codecleaning simattropts into unique class file readme layout starting attropts changes
