+++ src/ts/lib/GroupGlyphInterface.ts
@@ -4,8 +4,8 @@ import { SVGAttrOpts, SimulationAttrOpts } from "./DGLOsSVG";
 export interface GroupGlyph {
 	readonly groupType: string;
 	init(location: Selection<any, {}, any, {}>): Selection<any, {}, any, {}>;
-	initDraw(selection: Selection<any, {}, any, {}>, data: DynamicGraph, TimeStampIndex: number): Selection<any, {}, any, {}>;
-	updateDraw(selection: Selection<any, {}, any, {}>, attr: SVGAttrOpts, data: DynamicGraph, TimeStampIndex: number): Selection<any, {}, any, {}>;
+	initDraw(selection: Selection<any, {}, any, {}>): Selection<any, {}, any, {}>;
+	updateDraw(selection: Selection<any, {}, any, {}>, attr: SVGAttrOpts): Selection<any, {}, any, {}>;
 	transformTo(sourceG: Selection<any, {}, any, {}>, targetShape: GroupGlyph, targetG: Selection<any, {}, any, {}>): void;
 	draw(selection: Selection<any, {}, any, {}>, dGraph: DynamicGraph, TimeStampIndex: number, attr: SVGAttrOpts, noisePoints?: any, voronoiData?: any, enterExit?: any): void;
 }
